Backend/requirements.txt
python-dotenv
openai
langchain
langchain-core
langchain-huggingface
langchain-chroma
python-docx
sentence-transformers